// src/app/shared/model/Theme.ts
export interface Theme {
  name: string;
  headerBackground: string;
  headerTexte: string;
  backgroundGeneral: string;
  backgroundGrille: string;
  backgroundCelluleMorte: string;
  backgroundBorderCellule: string;
  backgroundCelluleVivante: string;
  backgroundBouton: string;
  boutonTexte: string;
  boutonHover: string
}